# 桌宠demo/新建文件夹/avatar-web-management/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AstralFox Market API",
    "version": "1.0.0",
    "description": "AI 桌面宠物市场 REST API。提供商品浏览、购买、评价、卖家管理等功能。所有端点（除公开端点外）需要 Bearer Token 认证。"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "本地开发" },
    { "url": "https://avatar.example.com", "description": "生产环境" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/settings/profile": {
      "get": {
        "tags": ["设置"],
        "summary": "获取用户资料",
        "responses": {
          "200": { "description": "用户资料" }
        }
      },
      "put": {
        "tags": ["设置"],
        "summary": "更新用户资料",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "bio": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "资料已更新" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}